Added Cart Model, and other random changes in random files
