+++ client/components/One-Product.js
@@ -3,7 +3,8 @@ import {Link} from 'react-router-dom'
 
 export const OneProduct = props => {
   return (
-    <div>
+    // className is for easy acces in css styling later
+    <div className="allProducts">
       <Link to={`/products/${props.product.id}`}>
         <h2>{props.product.name}</h2>
         <img src={props.product.imageUrl} />
